requirements.txt
scikit-learn==1.7.1
joblib==1.5.1
flask==3.1.0
numpy==2.3.2
scipy==1.16.1
requests==2.32.4
pytest==8.4.1
pytest-cov==6.2.1
flask-cors==6.0.1
pyodbc==5.2.0
python-dotenv==1.1.1
PyJWT==2.10.1

